backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
pandas==2.1.3
numpy==1.26.2
numpy-financial==1.0.0
requests==2.31.0
google-generativeai==0.8.3
groq==0.11.0
python-dotenv==1.0.0
scipy==1.11.4
httpx==0.25.2
aiohttp==3.9.1
